+++ tests/web/APIRest.php
@@ -42,6 +42,7 @@ use Computer;
 use Config;
 use Glpi\Tests\Api\Deprecated\Computer_SoftwareLicense;
 use Glpi\Tests\Api\Deprecated\Computer_SoftwareVersion;
+use Glpi\Tests\Api\Deprecated\ComputerAntivirus;
 use Glpi\Tests\Api\Deprecated\TicketFollowup;
 use GuzzleHttp;
 use Item_DeviceSimcard;
@@ -2332,6 +2333,7 @@ class APIRest extends atoum
             ['provider' => TicketFollowup::class],
             ['provider' => Computer_SoftwareVersion::class],
             ['provider' => Computer_SoftwareLicense::class],
+            ['provider' => ComputerAntivirus::class],
         ];
     }